Fix compile warnings from random key command
Reviewed-on: http://review.couchbase.org/30389
Tested-by: Chiyoung Seo <chiyoung@couchbase.com>

+++ src/stored-value.cc
@@ -629,8 +629,8 @@ Item *HashTable::getRandomKeyFromSlot(int slot) {
 
 Item* HashTable::getRandomKey(long rnd) {
     /* Try to locate a partition */
-    int start = rnd % size;
-    int curr = start;
+    size_t start = rnd % size;
+    size_t curr = start;
     Item *ret;
 
     do {
